requirements.txt

Django>=4.2,<5.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-filter>=24.2
drf-spectacular>=0.27.2
python-dateutil>=2.8.2
